# feat: Add comprehensive prerequisites system for Ubuntu 25.04 development environment