## 🚀 Prerequisites System for Ubuntu 25.04

This PR adds a comprehensive prerequisites system that transforms a fresh Ubuntu 25.04 system into a fully functional development environment.

### 🎯 Key Features

- **Modular Installation Scripts**: 8 specialized scripts for different components
- **Development Tools Suite**: Go, Node.js, Python, cloud CLIs, monitoring tools
- **Helper Scripts**: dev-manager, dev-check, system-monitor, setup-verification
- **System Automation**: One-command setup from fresh Ubuntu to full development environment
- **Environment Configuration**: Automatic PATH, aliases, and system optimizations

### 📁 New Files Added

-  - Comprehensive documentation
-  - Main orchestration script
- [0;34m[INFO][0m Updating package lists...
Reading package lists... - System updates and basic tools
- [0;35m[STEP][0m Starting libvirt setup...
[0;31m[ERROR][0m This script must be run as root (use sudo) - KVM/libvirt virtualization
- [0;35m[STEP][0m Starting Docker setup...
[0;31m[ERROR][0m This script must be run as root (use sudo) - Docker container runtime
- [0;35m[STEP][0m Starting Jenkins setup...
[0;31m[ERROR][0m This script must be run as root (use sudo) - Jenkins CI/CD server
- [0;35m[STEP][0m Starting Kubernetes tools setup...
[0;31m[ERROR][0m This script must be run as root (use sudo) - Kubernetes tools
- [0;35m[STEP][0m Starting Terraform setup...
[0;31m[ERROR][0m This script must be run as root (use sudo) - Terraform and IaC tools
- [0;34m[INFO][0m Starting Development Tools Setup v1.0.0
[0;34m[INFO][0m Starting development tools installation...
[0;35m[STEP][0m Updating package lists...
Hit:1 https://download.docker.com/linux/ubuntu plucky InRelease
Hit:2 http://security.ubuntu.com/ubuntu plucky-security InRelease
Hit:3 https://apt.releases.hashicorp.com plucky InRelease
Ign:4 https://pkg.jenkins.io/debian-stable binary/ InRelease
Hit:5 https://pkg.jenkins.io/debian-stable binary/ Release
Hit:6 http://ro.archive.ubuntu.com/ubuntu plucky InRelease
Hit:7 https://baltocdn.com/helm/stable/debian all InRelease
Hit:9 http://ro.archive.ubuntu.com/ubuntu plucky-updates InRelease
Hit:11 http://ro.archive.ubuntu.com/ubuntu plucky-backports InRelease
Hit:10 https://prod-cdn.packages.k8s.io/repositories/isv:/kubernetes:/core:/stable:/v1.33/deb  InRelease
Reading package lists...
[0;35m[STEP][0m Installing Go programming language...
[0;34m[INFO][0m Downloading Go 1.21.5...
[0;34m[INFO][0m Installing Go to /usr/local...
export PATH=$PATH:/usr/local/go/bin
[0;32m[SUCCESS][0m Go installed successfully: go version go1.21.5 linux/amd64
[0;35m[STEP][0m Installing Node.js and npm...
[38;5;79m2025-08-20 14:38:41 - Installing pre-requisites[0m
Hit:1 http://security.ubuntu.com/ubuntu plucky-security InRelease
Hit:2 https://apt.releases.hashicorp.com plucky InRelease
Hit:3 http://ro.archive.ubuntu.com/ubuntu plucky InRelease
Hit:4 https://download.docker.com/linux/ubuntu plucky InRelease
Hit:5 http://ro.archive.ubuntu.com/ubuntu plucky-updates InRelease
Hit:6 http://ro.archive.ubuntu.com/ubuntu plucky-backports InRelease
Hit:7 https://baltocdn.com/helm/stable/debian all InRelease
Ign:9 https://pkg.jenkins.io/debian-stable binary/ InRelease
Hit:10 https://pkg.jenkins.io/debian-stable binary/ Release
Hit:8 https://prod-cdn.packages.k8s.io/repositories/isv:/kubernetes:/core:/stable:/v1.33/deb  InRelease
Reading package lists...
Reading package lists...
Building dependency tree...
Reading state information...
apt-transport-https is already the newest version (3.0.0).
ca-certificates is already the newest version (20241223).
curl is already the newest version (8.12.1-3ubuntu1).
gnupg is already the newest version (2.4.4-2ubuntu23.1).
Solving dependencies...
0 upgraded, 0 newly installed, 0 to remove and 0 not upgraded.
Hit:1 https://apt.releases.hashicorp.com plucky InRelease
Hit:2 https://download.docker.com/linux/ubuntu plucky InRelease
Hit:3 http://ro.archive.ubuntu.com/ubuntu plucky InRelease
Get:4 https://deb.nodesource.com/node_20.x nodistro InRelease [12.1 kB]
Hit:5 http://ro.archive.ubuntu.com/ubuntu plucky-updates InRelease
Hit:6 https://baltocdn.com/helm/stable/debian all InRelease
Hit:7 http://ro.archive.ubuntu.com/ubuntu plucky-backports InRelease
Hit:8 http://security.ubuntu.com/ubuntu plucky-security InRelease
Get:9 https://deb.nodesource.com/node_20.x nodistro/main amd64 Packages [12.2 kB]
Ign:10 https://pkg.jenkins.io/debian-stable binary/ InRelease
Hit:12 https://pkg.jenkins.io/debian-stable binary/ Release
Hit:11 https://prod-cdn.packages.k8s.io/repositories/isv:/kubernetes:/core:/stable:/v1.33/deb  InRelease
Fetched 24.4 kB in 1s (46.4 kB/s)
Reading package lists...
[1;34m2025-08-20 14:38:45 - Repository configured successfully.[0m
[38;5;79m2025-08-20 14:38:45 - To install Node.js, run: apt-get install nodejs -y[0m
[38;5;79m2025-08-20 14:38:45 - You can use N|solid Runtime as a node.js alternative[0m
[1;32m2025-08-20 14:38:45 - To install N|solid Runtime, run: apt-get install nsolid -y 
[0m
Reading package lists...
Building dependency tree...
Reading state information...
Solving dependencies...
The following NEW packages will be installed:
  nodejs
0 upgraded, 1 newly installed, 0 to remove and 0 not upgraded.
Need to get 32.0 MB of archives.
After this operation, 199 MB of additional disk space will be used.
Get:1 https://deb.nodesource.com/node_20.x nodistro/main amd64 nodejs amd64 20.19.4-1nodesource1 [32.0 MB]
Fetched 32.0 MB in 1s (61.6 MB/s)
Selecting previously unselected package nodejs.
(Reading database ... (Reading database ... 5%(Reading database ... 10%(Reading database ... 15%(Reading database ... 20%(Reading database ... 25%(Reading database ... 30%(Reading database ... 35%(Reading database ... 40%(Reading database ... 45%(Reading database ... 50%(Reading database ... 55%(Reading database ... 60%(Reading database ... 65%(Reading database ... 70%(Reading database ... 75%(Reading database ... 80%(Reading database ... 85%(Reading database ... 90%(Reading database ... 95%(Reading database ... 100%(Reading database ... 190288 files and directories currently installed.)
Preparing to unpack .../nodejs_20.19.4-1nodesource1_amd64.deb ...
Unpacking nodejs (20.19.4-1nodesource1) ...
Setting up nodejs (20.19.4-1nodesource1) ...
Processing triggers for man-db (2.13.0-1) ...
[0;34m[INFO][0m Installing useful npm packages...

added 51 packages in 2s

5 packages are looking for funding
  run `npm fund` for details
[0;32m[SUCCESS][0m Node.js installed successfully: v20.19.4
[0;32m[SUCCESS][0m npm installed successfully: 10.8.2
[0;35m[STEP][0m Installing Python development tools...
[0;34m[INFO][0m Installing Python development packages...
Requirement already satisfied: requests in /usr/lib/python3/dist-packages (2.32.3)
Collecting beautifulsoup4
  Downloading beautifulsoup4-4.13.4-py3-none-any.whl.metadata (3.8 kB)
Collecting pandas
  Downloading pandas-2.3.1-cp313-cp313-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (91 kB)
Collecting numpy
  Downloading numpy-2.3.2-cp313-cp313-manylinux_2_27_x86_64.manylinux_2_28_x86_64.whl.metadata (62 kB)
Collecting matplotlib
  Downloading matplotlib-3.10.5-cp313-cp313-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata (11 kB)
Collecting jupyter
  Downloading jupyter-1.1.1-py2.py3-none-any.whl.metadata (2.0 kB)
Collecting ipython
  Downloading ipython-9.4.0-py3-none-any.whl.metadata (4.4 kB)
Collecting pytest
  Downloading pytest-8.4.1-py3-none-any.whl.metadata (7.7 kB)
Requirement already satisfied: black in /usr/lib/python3/dist-packages (25.1.0)
Collecting flake8
  Downloading flake8-7.3.0-py2.py3-none-any.whl.metadata (3.8 kB)
Collecting mypy
  Downloading mypy-1.17.1-cp313-cp313-manylinux2014_x86_64.manylinux_2_17_x86_64.manylinux_2_28_x86_64.whl.metadata (2.2 kB)
Collecting pre-commit
  Downloading pre_commit-4.3.0-py2.py3-none-any.whl.metadata (1.2 kB)
Requirement already satisfied: chardet>=3.0.2 in /usr/lib/python3/dist-packages (from requests) (5.2.0)
Requirement already satisfied: idna<4,>=2.5 in /usr/lib/python3/dist-packages (from requests) (3.10)
Requirement already satisfied: urllib3<3,>=1.21.1 in /usr/lib/python3/dist-packages (from requests) (2.3.0)
Requirement already satisfied: certifi>=2017.4.17 in /usr/lib/python3/dist-packages (from requests) (2025.1.31)
Collecting soupsieve>1.2 (from beautifulsoup4)
  Downloading soupsieve-2.7-py3-none-any.whl.metadata (4.6 kB)
Requirement already satisfied: typing-extensions>=4.0.0 in /usr/lib/python3/dist-packages (from beautifulsoup4) (4.12.2)
Requirement already satisfied: python-dateutil>=2.8.2 in /usr/lib/python3/dist-packages (from pandas) (2.9.0)
Requirement already satisfied: pytz>=2020.1 in /usr/lib/python3/dist-packages (from pandas) (2025.1)
Collecting tzdata>=2022.7 (from pandas)
  Downloading tzdata-2025.2-py2.py3-none-any.whl.metadata (1.4 kB)
Collecting contourpy>=1.0.1 (from matplotlib)
  Downloading contourpy-1.3.3-cp313-cp313-manylinux_2_27_x86_64.manylinux_2_28_x86_64.whl.metadata (5.5 kB)
Collecting cycler>=0.10 (from matplotlib)
  Downloading cycler-0.12.1-py3-none-any.whl.metadata (3.8 kB)
Collecting fonttools>=4.22.0 (from matplotlib)
  Downloading fonttools-4.59.1-cp313-cp313-manylinux1_x86_64.manylinux2014_x86_64.manylinux_2_17_x86_64.manylinux_2_5_x86_64.whl.metadata (108 kB)
Collecting kiwisolver>=1.3.1 (from matplotlib)
  Downloading kiwisolver-1.4.9-cp313-cp313-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata (6.3 kB)
Requirement already satisfied: packaging>=20.0 in /usr/lib/python3/dist-packages (from matplotlib) (24.2)
Requirement already satisfied: pillow>=8 in /usr/lib/python3/dist-packages (from matplotlib) (11.1.0)
Requirement already satisfied: pyparsing>=2.3.1 in /usr/lib/python3/dist-packages (from matplotlib) (3.1.2)
Collecting notebook (from jupyter)
  Downloading notebook-7.4.5-py3-none-any.whl.metadata (10 kB)
Collecting jupyter-console (from jupyter)
  Downloading jupyter_console-6.6.3-py3-none-any.whl.metadata (5.8 kB)
Collecting nbconvert (from jupyter)
  Downloading nbconvert-7.16.6-py3-none-any.whl.metadata (8.5 kB)
Collecting ipykernel (from jupyter)
  Downloading ipykernel-6.30.1-py3-none-any.whl.metadata (6.2 kB)
Collecting ipywidgets (from jupyter)
  Downloading ipywidgets-8.1.7-py3-none-any.whl.metadata (2.4 kB)
Collecting jupyterlab (from jupyter)
  Downloading jupyterlab-4.4.6-py3-none-any.whl.metadata (16 kB)
Requirement already satisfied: decorator in /usr/lib/python3/dist-packages (from ipython) (5.1.1)
Collecting ipython-pygments-lexers (from ipython)
  Downloading ipython_pygments_lexers-1.1.1-py3-none-any.whl.metadata (1.1 kB)
Collecting jedi>=0.16 (from ipython)
  Downloading jedi-0.19.2-py2.py3-none-any.whl.metadata (22 kB)
Collecting matplotlib-inline (from ipython)
  Downloading matplotlib_inline-0.1.7-py3-none-any.whl.metadata (3.9 kB)
Requirement already satisfied: pexpect>4.3 in /usr/lib/python3/dist-packages (from ipython) (4.9.0)
Collecting prompt_toolkit<3.1.0,>=3.0.41 (from ipython)
  Downloading prompt_toolkit-3.0.51-py3-none-any.whl.metadata (6.4 kB)
Requirement already satisfied: pygments>=2.4.0 in /usr/lib/python3/dist-packages (from ipython) (2.18.0)
Collecting stack_data (from ipython)
  Downloading stack_data-0.6.3-py3-none-any.whl.metadata (18 kB)
Collecting traitlets>=5.13.0 (from ipython)
  Downloading traitlets-5.14.3-py3-none-any.whl.metadata (10 kB)
Collecting iniconfig>=1 (from pytest)
  Downloading iniconfig-2.1.0-py3-none-any.whl.metadata (2.7 kB)
Collecting pluggy<2,>=1.5 (from pytest)
  Downloading pluggy-1.6.0-py3-none-any.whl.metadata (4.8 kB)
Requirement already satisfied: click>=8.0.0 in /usr/lib/python3/dist-packages (from black) (8.1.8)
Requirement already satisfied: mypy-extensions>=0.4.3 in /usr/lib/python3/dist-packages (from black) (1.0.0)
Requirement already satisfied: pathspec>=0.9.0 in /usr/lib/python3/dist-packages (from black) (0.12.1)
Requirement already satisfied: platformdirs>=2 in /usr/lib/python3/dist-packages (from black) (4.3.6)
Collecting mccabe<0.8.0,>=0.7.0 (from flake8)
  Downloading mccabe-0.7.0-py2.py3-none-any.whl.metadata (5.0 kB)
Collecting pycodestyle<2.15.0,>=2.14.0 (from flake8)
  Downloading pycodestyle-2.14.0-py2.py3-none-any.whl.metadata (4.5 kB)
Collecting pyflakes<3.5.0,>=3.4.0 (from flake8)
  Downloading pyflakes-3.4.0-py2.py3-none-any.whl.metadata (3.5 kB)
Collecting cfgv>=2.0.0 (from pre-commit)
  Downloading cfgv-3.4.0-py2.py3-none-any.whl.metadata (8.5 kB)
Collecting identify>=1.0.0 (from pre-commit)
  Downloading identify-2.6.13-py2.py3-none-any.whl.metadata (4.4 kB)
Collecting nodeenv>=0.11.1 (from pre-commit)
  Downloading nodeenv-1.9.1-py2.py3-none-any.whl.metadata (21 kB)
Requirement already satisfied: pyyaml>=5.1 in /usr/lib/python3/dist-packages (from pre-commit) (6.0.2)
Collecting virtualenv>=20.10.0 (from pre-commit)
  Downloading virtualenv-20.34.0-py3-none-any.whl.metadata (4.6 kB)
Collecting parso<0.9.0,>=0.8.4 (from jedi>=0.16->ipython)
  Downloading parso-0.8.4-py2.py3-none-any.whl.metadata (7.7 kB)
Requirement already satisfied: ptyprocess>=0.5 in /usr/lib/python3/dist-packages (from pexpect>4.3->ipython) (0.7.0)
Collecting wcwidth (from prompt_toolkit<3.1.0,>=3.0.41->ipython)
  Downloading wcwidth-0.2.13-py2.py3-none-any.whl.metadata (14 kB)
Collecting distlib<1,>=0.3.7 (from virtualenv>=20.10.0->pre-commit)
  Downloading distlib-0.4.0-py2.py3-none-any.whl.metadata (5.2 kB)
Requirement already satisfied: filelock<4,>=3.12.2 in /usr/lib/python3/dist-packages (from virtualenv>=20.10.0->pre-commit) (3.18.0)
Collecting comm>=0.1.1 (from ipykernel->jupyter)
  Downloading comm-0.2.3-py3-none-any.whl.metadata (3.7 kB)
Collecting debugpy>=1.6.5 (from ipykernel->jupyter)
  Downloading debugpy-1.8.16-cp313-cp313-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (1.3 kB)
Collecting jupyter-client>=8.0.0 (from ipykernel->jupyter)
  Downloading jupyter_client-8.6.3-py3-none-any.whl.metadata (8.3 kB)
Collecting jupyter-core!=5.0.*,>=4.12 (from ipykernel->jupyter)
  Downloading jupyter_core-5.8.1-py3-none-any.whl.metadata (1.6 kB)
Collecting nest-asyncio>=1.4 (from ipykernel->jupyter)
  Downloading nest_asyncio-1.6.0-py3-none-any.whl.metadata (2.8 kB)
Collecting psutil>=5.7 (from ipykernel->jupyter)
  Downloading psutil-7.0.0-cp36-abi3-manylinux_2_12_x86_64.manylinux2010_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (22 kB)
Collecting pyzmq>=25 (from ipykernel->jupyter)
  Downloading pyzmq-27.0.1-cp312-abi3-manylinux_2_26_x86_64.manylinux_2_28_x86_64.whl.metadata (6.0 kB)
Collecting tornado>=6.2 (from ipykernel->jupyter)
  Downloading tornado-6.5.2-cp39-abi3-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (2.8 kB)
Collecting widgetsnbextension~=4.0.14 (from ipywidgets->jupyter)
  Downloading widgetsnbextension-4.0.14-py3-none-any.whl.metadata (1.6 kB)
Collecting jupyterlab_widgets~=3.0.15 (from ipywidgets->jupyter)
  Downloading jupyterlab_widgets-3.0.15-py3-none-any.whl.metadata (20 kB)
Collecting async-lru>=1.0.0 (from jupyterlab->jupyter)
  Downloading async_lru-2.0.5-py3-none-any.whl.metadata (4.5 kB)
Collecting httpx<1,>=0.25.0 (from jupyterlab->jupyter)
  Downloading httpx-0.28.1-py3-none-any.whl.metadata (7.1 kB)
Requirement already satisfied: jinja2>=3.0.3 in /usr/lib/python3/dist-packages (from jupyterlab->jupyter) (3.1.5)
Collecting jupyter-lsp>=2.0.0 (from jupyterlab->jupyter)
  Downloading jupyter_lsp-2.2.6-py3-none-any.whl.metadata (1.8 kB)
Collecting jupyter-server<3,>=2.4.0 (from jupyterlab->jupyter)
  Downloading jupyter_server-2.16.0-py3-none-any.whl.metadata (8.5 kB)
Collecting jupyterlab-server<3,>=2.27.1 (from jupyterlab->jupyter)
  Downloading jupyterlab_server-2.27.3-py3-none-any.whl.metadata (5.9 kB)
Collecting notebook-shim>=0.2 (from jupyterlab->jupyter)
  Downloading notebook_shim-0.2.4-py3-none-any.whl.metadata (4.0 kB)
Collecting setuptools>=41.1.0 (from jupyterlab->jupyter)
  Downloading setuptools-80.9.0-py3-none-any.whl.metadata (6.6 kB)
Collecting bleach!=5.0.0 (from bleach[css]!=5.0.0->nbconvert->jupyter)
  Downloading bleach-6.2.0-py3-none-any.whl.metadata (30 kB)
Collecting defusedxml (from nbconvert->jupyter)
  Downloading defusedxml-0.7.1-py2.py3-none-any.whl.metadata (32 kB)
Collecting jupyterlab-pygments (from nbconvert->jupyter)
  Downloading jupyterlab_pygments-0.3.0-py3-none-any.whl.metadata (4.4 kB)
Requirement already satisfied: markupsafe>=2.0 in /usr/lib/python3/dist-packages (from nbconvert->jupyter) (2.1.5)
Collecting mistune<4,>=2.0.3 (from nbconvert->jupyter)
  Downloading mistune-3.1.3-py3-none-any.whl.metadata (1.8 kB)
Collecting nbclient>=0.5.0 (from nbconvert->jupyter)
  Downloading nbclient-0.10.2-py3-none-any.whl.metadata (8.3 kB)
Collecting nbformat>=5.7 (from nbconvert->jupyter)
  Downloading nbformat-5.10.4-py3-none-any.whl.metadata (3.6 kB)
Collecting pandocfilters>=1.4.1 (from nbconvert->jupyter)
  Downloading pandocfilters-1.5.1-py2.py3-none-any.whl.metadata (9.0 kB)
Collecting executing>=1.2.0 (from stack_data->ipython)
  Downloading executing-2.2.0-py2.py3-none-any.whl.metadata (8.9 kB)
Collecting asttokens>=2.1.0 (from stack_data->ipython)
  Downloading asttokens-3.0.0-py3-none-any.whl.metadata (4.7 kB)
Collecting pure-eval (from stack_data->ipython)
  Downloading pure_eval-0.2.3-py3-none-any.whl.metadata (6.3 kB)
Collecting webencodings (from bleach!=5.0.0->bleach[css]!=5.0.0->nbconvert->jupyter)
  Downloading webencodings-0.5.1-py2.py3-none-any.whl.metadata (2.1 kB)
Collecting tinycss2<1.5,>=1.1.0 (from bleach[css]!=5.0.0->nbconvert->jupyter)
  Downloading tinycss2-1.4.0-py3-none-any.whl.metadata (3.0 kB)
Collecting anyio (from httpx<1,>=0.25.0->jupyterlab->jupyter)
  Downloading anyio-4.10.0-py3-none-any.whl.metadata (4.0 kB)
Collecting httpcore==1.* (from httpx<1,>=0.25.0->jupyterlab->jupyter)
  Downloading httpcore-1.0.9-py3-none-any.whl.metadata (21 kB)
Collecting h11>=0.16 (from httpcore==1.*->httpx<1,>=0.25.0->jupyterlab->jupyter)
  Downloading h11-0.16.0-py3-none-any.whl.metadata (8.3 kB)
Collecting argon2-cffi>=21.1 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading argon2_cffi-25.1.0-py3-none-any.whl.metadata (4.1 kB)
Collecting jupyter-events>=0.11.0 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading jupyter_events-0.12.0-py3-none-any.whl.metadata (5.8 kB)
Collecting jupyter-server-terminals>=0.4.4 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading jupyter_server_terminals-0.5.3-py3-none-any.whl.metadata (5.6 kB)
Collecting overrides>=5.0 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading overrides-7.7.0-py3-none-any.whl.metadata (5.8 kB)
Collecting prometheus-client>=0.9 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading prometheus_client-0.22.1-py3-none-any.whl.metadata (1.9 kB)
Collecting send2trash>=1.8.2 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading Send2Trash-1.8.3-py3-none-any.whl.metadata (4.0 kB)
Collecting terminado>=0.8.3 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading terminado-0.18.1-py3-none-any.whl.metadata (5.8 kB)
Collecting websocket-client>=1.7 (from jupyter-server<3,>=2.4.0->jupyterlab->jupyter)
  Downloading websocket_client-1.8.0-py3-none-any.whl.metadata (8.0 kB)
Requirement already satisfied: babel>=2.10 in /usr/lib/python3/dist-packages (from jupyterlab-server<3,>=2.27.1->jupyterlab->jupyter) (2.17.0)
Collecting json5>=0.9.0 (from jupyterlab-server<3,>=2.27.1->jupyterlab->jupyter)
  Downloading json5-0.12.1-py3-none-any.whl.metadata (36 kB)
Requirement already satisfied: jsonschema>=4.18.0 in /usr/lib/python3/dist-packages (from jupyterlab-server<3,>=2.27.1->jupyterlab->jupyter) (4.19.2)
Collecting fastjsonschema>=2.15 (from nbformat>=5.7->nbconvert->jupyter)
  Downloading fastjsonschema-2.21.2-py3-none-any.whl.metadata (2.3 kB)
…